Cross-validated AuditorFitter from a Learner

CVLearnerAuditorFitter returns the cross-validated predictions instead of the in-sample predictions.

Available data is cut into complementary subsets (folds). For each subset out-of-sample predictions are received by training a model on all other subsets and predicting afterwards on the left-out subset.

Value

AuditorFitter

list with items

corr: pseudo-correlation between residuals and learner prediction.

l: the trained learner.

Functions

CVTreeAuditorFitter: Cross-Validated auditor based on rpart

CVRidgeAuditorFitter: Cross-Validated auditor based on glmnet

See also

Super class

mcboost::AuditorFitter -> CVLearnerAuditorFitter

Public fields

learner

CVLearnerPredictor
Learner used for fitting residuals.

Methods

Public methods

CVLearnerAuditorFitter$new()

CVLearnerAuditorFitter$fit()

CVLearnerAuditorFitter$clone()

Inherited methods

mcboost::AuditorFitter$fit_to_resid()

Method new()

Define a CVAuditorFitter from a learner. Available instantiations:
CVTreeAuditorFitter (rpart) and CVRidgeAuditorFitter (glmnet). See mlr3pipelines::PipeOpLearnerCV for more information on cross-validated learners.

Usage

CVLearnerAuditorFitter$new(learner, folds = 3L)

Arguments

learner

mlr3::Learner
Regression Learner to use.

folds

integer
Number of folds to use for PipeOpLearnerCV. Defaults to 3.

Method fit()

Fit the cross-validated learner and compute correlation

Usage

CVLearnerAuditorFitter$fit(data, resid, mask)

Arguments

data

data.table
Features.

resid

numeric
Residuals (of same length as data).

mask

integer
Mask applied to the data. Only used for SubgroupAuditorFitter.

Method clone()

The objects of this class are cloneable with this method.

Usage

CVLearnerAuditorFitter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

mcboost::AuditorFitter -> mcboost::CVLearnerAuditorFitter -> CVTreeAuditorFitter

Methods

Public methods

CVTreeAuditorFitter$new()

CVTreeAuditorFitter$clone()

Inherited methods

mcboost::AuditorFitter$fit_to_resid()

mcboost::CVLearnerAuditorFitter$fit()

Method new()

Define a cross-validated AuditorFitter from a rpart learner See mlr3pipelines::PipeOpLearnerCV for more information on cross-validated learners.

Usage

CVTreeAuditorFitter$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

CVTreeAuditorFitter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

mcboost::AuditorFitter -> mcboost::CVLearnerAuditorFitter -> CVRidgeAuditorFitter

Methods

Public methods

CVRidgeAuditorFitter$new()

CVRidgeAuditorFitter$clone()

Inherited methods

mcboost::AuditorFitter$fit_to_resid()

mcboost::CVLearnerAuditorFitter$fit()

Method new()

Define a cross-validated AuditorFitter from a glmnet learner. See mlr3pipelines::PipeOpLearnerCV for more information on cross-validated learners.

Usage

CVRidgeAuditorFitter$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

CVRidgeAuditorFitter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.