Multi-Calibration Boosting

Implements Multi-Calibration Boosting by Hebert-Johnson et al. (2018) and Multi-Accuracy Boosting by Kim et al. (2019) for the multi-calibration of a machine learning model's prediction for survival models. Multi-Calibration works best in scenarios where the underlying data & labels are unbiased but a bias is introduced within the algorithm's fitting procedure. This is often the case, e.g. when an algorithm fits a majority population while ignoring or under-fitting minority populations.
Expects initial models that predict probobilities (between 0 and 1) for different time points. The method defaults to Multi-Accuracy Boosting as described in Kim et al. (2019). In order to obtain behaviour as described in Hebert-Johnson et al. (2018) set multiplicative=FALSE and num_buckets to 10.

Hebert-Johnson et al., 2018. Multicalibration: Calibration for the (Computationally-Identifiable) Masses. Proceedings of the 35th International Conference on Machine Learning, PMLR 80:1939-1948. https://proceedings.mlr.press/v80/hebert-johnson18a.html.

Kim et al., 2019. Multiaccuracy: Black-Box Post-Processing for Fairness in Classification. Proceedings of the 2019 AAAI/ACM Conference on AI, Ethics, and Society (AIES '19). Association for Computing Machinery, New York, NY, USA, 247–254. https://dl.acm.org/doi/10.1145/3306618.3314287

Super class

mcboost::MCBoost -> MCBoostSurv

Public fields

max_iter

integer
The maximum number of iterations of the multi-calibration/multi-accuracy method.

alpha

numeric
Accuracy parameter that determines the stopping condition.

eta

numeric
Parameter for multiplicative weight update (step size).

num_buckets

integer
The number of buckets to split into in addition to using the whole sample.

bucket_strategy

character
Currently only supports "simple", even split along probabilities. Only relevant for num_buckets > 1.

rebucket

logical
Should buckets be re-calculated at each iteration?

eval_fulldata

logical
Should auditor be evaluated on the full data?

partition

logical
True/False flag for whether to split up predictions by their "partition" (e.g., predictions less than 0.5 and predictions greater than 0.5).

multiplicative

logical
Specifies the strategy for updating the weights (multiplicative weight vs additive).

iter_sampling

character
Specifies the strategy to sample the validation data for each iteration.

auditor_fitter

AuditorFitter
Specifies the type of model used to fit the residuals.

predictor

function
Initial predictor function.

iter_models

list
Cumulative list of fitted models.

iter_partitions

list
Cumulative list of data partitions for models.

iter_corr

list
Auditor correlation in each iteration.

auditor_effects

list
Auditor effect in each iteration.

time_points

integer
Times included in the prediction (columnames)

time_buckets

integer
The number of buckets to split the time points (columns) of the prediction.

bucket_strategies

character
Possible bucket_strategies in McBoostSurv. Only relevant for time_buckets > 1. even_splits: split buckets evenly quantiles: split buckets by quantiles

bucket_aggregation

function
If not NULL, predictions are not selected by time/probability, but by time/individual. Individuals are selected by aggregated value per individual (e.g. mean). Only relevant for time_buckets > 1.

max_time_quantile

double
Time quantile which should be evaluated and multicalibrated. Similar to a 75%-Integrated Brier Score.

time_points_eval

integer
Vector of time_points that should be evaluated.

loss

character
Loss function which is optimized during boosting. censored_brier: censored version of the integrated brier score brier: uncensored version of the integrated brier score censored_brier_proper: proper version of the censored version of the integrated brier score For more details, we are referring to https://mlr3proba.mlr-org.com/reference/mlr_measures_surv.graf.html.

Methods

Public methods

MCBoostSurv$new()

MCBoostSurv$clone()

Inherited methods

mcboost::MCBoost$auditor_effect()

mcboost::MCBoost$multicalibrate()

mcboost::MCBoost$predict_probs()

mcboost::MCBoost$print()

Method new()

Initialize a multi-calibration instance.

Usage

MCBoostSurv$new(
  max_iter = 25,
  alpha = 1e-04,
  eta = 0.1,
  num_buckets = 1,
  partition = ifelse(num_buckets > 1, TRUE, FALSE),
  time_buckets = 2L,
  max_time_quantile = 1,
  bucket_strategy = "even_splits",
  bucket_aggregation = NULL,
  rebucket = FALSE,
  eval_fulldata = FALSE,
  multiplicative = TRUE,
  auditor_fitter = "RidgeAuditorFitter",
  subpops = NULL,
  default_model_class = LearnerSurvKaplan,
  init_predictor = NULL,
  loss = "censored_brier",
  iter_sampling = "none"
)

Arguments

max_iter

integer
The maximum number of iterations of the multi-calibration/multi-accuracy method. Default 5L.

alpha

numeric
Accuracy parameter that determines the stopping condition. Default 1e-4.

eta

numeric
Parameter for multiplicative weight update (step size). Default 1.0.

num_buckets

integer
The number of buckets to split into in addition to using the whole sample. Default 2L.

partition

logical
True/False flag for whether to split up predictions by their "partition" (e.g., predictions less than 0.5 and predictions greater than 0.5). Defaults to TRUE (multi-accuracy boosting).

time_buckets

integer
The number of buckets to split the time points (columns) of the prediction.

max_time_quantile

double
Time quantile which should be evaluated and multicalibrated. Can be used to perform multi-calibration only up to the max_time_quantile percent of timepoints. Initialized to 1.

bucket_strategy

character
Bucketstragy for bucketing. even_splits: split buckets evenly quantiles: split buckets by quantiles

bucket_aggregation

function
If not NULL, predictions are not selected by time/probability, but by time/individual. Individuals are selected by aggregated value per individual (e.g. mean). Only relevant for time_buckets > 1.

rebucket

logical
Should buckets be re-done at each iteration? Default FALSE.

eval_fulldata

logical
Should the auditor be evaluated on the full data or on the respective bucket for determining the stopping criterion? Default FALSE, auditor is only evaluated on the bucket. This setting keeps the implementation closer to the Algorithm proposed in the corresponding multi-accuracy paper (Kim et al., 2019) where auditor effects are computed across the full sample (i.e. eval_fulldata = TRUE).

multiplicative

logical
Specifies the strategy for updating the weights (multiplicative weight vs additive). Defaults to TRUE (multi-accuracy boosting). Set to FALSE for multi-calibration.

auditor_fitter

AuditorFitter|character|mlr3::Learner
Specifies the type of model used to fit the residuals. The default is RidgeAuditorFitter. Can be a character, the name of a AuditorFitter, a mlr3::Learner that is then auto-converted into a LearnerAuditorFitter or a custom AuditorFitter.

subpops

list
Specifies a collection of characteristic attributes and the values they take to define subpopulations e.g. list(age = c('20-29','30-39','40+'), nJobs = c(0,1,2,'3+'), ,..).

default_model_class

Predictor
The class of the model that should be used as the init predictor model if init_predictor is not specified. Defaults to ConstantPredictor which predicts a constant value.

init_predictor

function|mlr3::Learner
The initial predictor function to use (i.e., if the user has a pretrained model). If a mlr3 Learner is passed, it will be autoconverted using mlr3_init_predictor. This requires the mlr3::Learner to be trained.

loss

character
#' Loss function which is optimized during boosting. censored_brier: censored version of the integrated brier score brier: uncensored version of the integrated brier score censored_brier_proper: proper version of the censored version of the integrated brier score For more details, we are referring to https://mlr3proba.mlr-org.com/reference/mlr_measures_surv.graf.html.

iter_sampling

character
How to sample the validation data for each iteration? Can be bootstrap, split or none.
"split" splits the data into max_iter parts and validates on each sample in each iteration.
"bootstrap" uses a new bootstrap sample in each iteration.
"none" uses the same dataset in each iteration.

Method clone()

The objects of this class are cloneable with this method.

Usage

MCBoostSurv$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.