mcboost: Multi-Calibration Boosting

Implements 'Multi-Calibration Boosting' (2018) <https://proceedings.mlr.press/v80/hebert-johnson18a.html> and 'Multi-Accuracy Boosting' (2019) <arXiv:1805.12317> for the multi-calibration of a machine learning model's prediction. 'MCBoost' updates predictions for sub-groups in an iterative fashion in order to mitigate biases like poor calibration or large accuracy differences across subgroups. Multi-Calibration works best in scenarios where the underlying data & labels are unbiased, but resulting models are. This is often the case, e.g. when an algorithm fits a majority population while ignoring or under-fitting minority populations.

References

Kim et al., 2019: Multiaccuracy: Black-Box Post-Processing for Fairness in Classification. Hebert-Johnson et al., 2018: Multicalibration: Calibration for the (Computationally-Identifiable) Masses.

See also

Author

Maintainer: Florian Pfisterer pfistererf@googlemail.com (ORCID)

Other contributors:

Susanne Dandl susanne.dandl@stat.uni-muenchen.de (ORCID) [contributor]

Christoph Kern c.kern@uni-mannheim.de (ORCID) [contributor]

Bernd Bischl bernd_bischl@gmx.net (ORCID) [contributor]