Multi-Calibrate a Learner's Prediction (Survival Model)

Post-process a survival learner prediction using multi-calibration. For more details, please refer to https://arxiv.org/pdf/1805.12317.pdf (Kim et al. 2018) or the help for MCBoostSurv. If no init_predictor is provided, the preceding learner's predictions corresponding to the prediction slot are used as an initial predictor for MCBoostSurv.

Format

R6Class inheriting from mlr3pipelines::PipeOp.

Construction

PipeOpMCBoostSurv$new(id = "mcboostsurv", param_vals = list())

id :: character(1) Identifier of the resulting object, default "threshold".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. See MCBoostSurv for a comprehensive description of all hyperparameters.

Input and Output Channels

During training, the input and output are "data" and "prediction", two TaskSurv. A PredictionSurv is required as input and returned as output during prediction.

State

The $state is a MCBoostSurv Object as obtained from MCBoostSurv$new().

Parameters

max_iter :: integer
A integer specifying the number of multi-calibration rounds. Defaults to 5.

Fields

Only fields inherited from mlr3pipelines::PipeOp.

Methods

Only methods inherited from mlr3pipelines::PipeOp.

See also

Super class

mlr3pipelines::PipeOp -> PipeOpMCBoostSurv

Active bindings

predict_type

Predict type of the PipeOp.

Methods

Public methods

PipeOpMCBoostSurv$new()

PipeOpMCBoostSurv$clone()

Inherited methods

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Initialize a Multi-Calibration PipeOp (for Survival).

Usage

PipeOpMCBoostSurv$new(id = "mcboostsurv", param_vals = list())

Arguments

id

character
The PipeOp's id. Defaults to "mcboostsurv".

param_vals

list
List of hyperparameters for the PipeOp.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpMCBoostSurv$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(mlr3pipelines)
# Attention: gunion inputs have to be in the correct order for now.
if (FALSE) {
gr = gunion(list(
  "data" = po("nop"),
  "prediction" = po("learner_pred", lrn("surv.ranger"))
)) %>>%
  PipeOpMCBoostSurv$new()
tsk = tsk("rats")
tid = sample(1:300, 100)
gr$train(tsk$clone()$filter(tid))
gr$predict(tsk$clone()$filter(setdiff(1:300, tid)))
}