Multi-Calibrate a Learner's Prediction

Post-process a learner prediction using multi-calibration. For more details, please refer to https://arxiv.org/pdf/1805.12317.pdf (Kim et al. 2018) or the help for MCBoost. If no init_predictor is provided, the preceding learner's predictions corresponding to the prediction slot are used as an initial predictor for MCBoost.

Format

R6Class inheriting from mlr3pipelines::PipeOp.

Construction

PipeOpMCBoost$new(id = "mcboost", param_vals = list())

id :: character(1) Identifier of the resulting object, default "threshold".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. See MCBoost for a comprehensive description of all hyperparameters.

Input and Output Channels

During training, the input and output are "data" and "prediction", two TaskClassif. A PredictionClassif is required as input and returned as output during prediction.

State

The $state is a MCBoost Object as obtained from MCBoost$new().

Parameters

max_iter :: integer
A integer specifying the number of multi-calibration rounds. Defaults to 5.

Fields

Only fields inherited from mlr3pipelines::PipeOp.

Methods

Only methods inherited from mlr3pipelines::PipeOp.

See also

Super class

mlr3pipelines::PipeOp -> PipeOpMCBoost

Active bindings

predict_type

Predict type of the PipeOp.

Methods

Public methods

PipeOpMCBoost$new()

PipeOpMCBoost$clone()

Inherited methods

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Initialize a Multi-Calibration PipeOp.

Usage

PipeOpMCBoost$new(id = "mcboost", param_vals = list())

Arguments

id

character
The PipeOp's id. Defaults to "mcboost".

param_vals

list
List of hyperparameters for the PipeOp.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpMCBoost$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(mlr3pipelines)
# Attention: gunion inputs have to be in the correct order for now.
 gr = gunion(list(
   "data" = po("nop"),
   "prediction" = po("learner_cv", lrn("classif.rpart"))
  )) %>>%
  PipeOpMCBoost$new()
tsk = tsk("sonar")
tid = sample(1:208, 108)
gr$train(tsk$clone()$filter(tid))

#> $mcboost.output
#> NULL
#> 

gr$predict(tsk$clone()$filter(setdiff(1:208, tid)))

#> $mcboost.output
#> <PredictionClassif> for 100 observations:
#>     row_ids truth response       prob.M     prob.R
#>           1     R        R 9.627347e-05 0.99990373
#>           5     R        M 9.066802e-01 0.09331983
#>           6     R        R 8.532412e-05 0.99991468
#> ---                                               
#>         205     M        R 9.669045e-05 0.99990331
#>         206     M        R 9.208786e-05 0.99990791
#>         207     M        M 1.000000e+00 0.00000000
#>